+++ AITest/Models/Layers/HiddenLayer.cs
@@ -7,16 +7,16 @@ public class HiddenLayer(int numberOfNeurons, int numberOfPreviousNeurons, Neuro
 {
     public override void Recognize(NeuralNetwork? network, Layer? nextLayer)
     {
-        var hiddenOut = new double[Neurons.Length];
+        var hiddenOut = new decimal[Neurons.Length];
         for (var i = 0; i < Neurons.Length; ++i)
             hiddenOut[i] = Neurons[i].Output;
         if (nextLayer != null) 
             nextLayer.Data = hiddenOut;
     }
 
-    public override double[]? BackwardPass(double[] gradientsSums)
+    public override decimal[]? BackwardPass(decimal[] gradientsSums)
     {
-        double[]? gradientSum = null;
+        decimal[]? gradientSum = null;
 
         for (var i = 0; i < NumberOfNeurons; ++i)
         for (var n = 0; n < NumberOfPreviousNeurons; ++n)
